refactor: moved directory and deleted unused components
