+++ app/(auth)/_components/SidePicture.tsx
